PCNT Interrupt Processing
In the previous version, it was found that the sensor HI inputs did not produce the correct values.  Typically they were found to be wrong by several 10s of counts.

After investigation, it was found that the software in "Final Push Before Live Fire" was incorrect since it only returned the counts on the NORTH_LOW register and was wrong.  Changing to the correct registers gave the correct, but (undesirable) value.

After a bit of investigation, it was found that the interrupt disable instruction was being executed in MANY instruction cycles.  This patch simplifies the interrupt disable by simply hammering the interrupt enable from within the interrupt service routine.

This patch has yet to be tested with live fire.